Update 01-mutations.md
Replaced the reference to ASP.NET Core MVC with the EntityGraphQL section on DI. Even though MVC uses a similar pattern, there are differences. For example, the ASP.NET Core MVC say to use FromServices, but assembly Microsoft.AspNetCore.Mvc, which hosts `FromServicesAttribute`, may not even be referenced, in a EntityGraphQL project that's not using ASP.NET Core.